@@ -11,48 +11,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.0",
-    "@angular/common": "~9.1.0",
-    "@angular/compiler": "~9.1.0",
-    "@angular/core": "~9.1.0",
-    "@angular/forms": "~9.1.0",
-    "@angular/localize": "^9.1.0",
-    "@angular/platform-browser": "~9.1.0",
-    "@angular/platform-browser-dynamic": "~9.1.0",
-    "@angular/router": "~9.1.0",
-    "@ng-bootstrap/ng-bootstrap": "^5.1.4",
-    "ag-grid-angular": "^23.0.3",
-    "ag-grid-community": "^23.0.2",
-    "bootstrap": "^3.4.1",
-    "core-js": "^3.6.1",
-    "file-saver": "^2.0.2",
-    "jquery": ">=3.5.0",
-    "ngx-bootstrap": "^5.3.2",
-    "popper.js": "^1.16.0",
-    "rxjs": "~6.5.5",
-    "rxjs-compat": "^6.5.4",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "@angular/animations": "^14.2.7",
+    "@angular/common": "~14.2.7",
+    "@angular/compiler": "~14.2.7",
+    "@angular/core": "~14.2.7",
+    "@angular/forms": "~14.2.7",
+    "@angular/localize": "^14.2.7",
+    "@angular/platform-browser": "~14.2.7",
+    "@angular/platform-browser-dynamic": "~14.2.7",
+    "@angular/router": "~14.2.7",
+    "@ng-bootstrap/ng-bootstrap": "^13.1.0",
+    "ag-grid-angular": "^28.2.0",
+    "ag-grid-community": "^28.2.0",
+    "bootstrap": "^5.2.2",
+    "core-js": "^3.26.0",
+    "file-saver": "^2.0.5",
+    "jquery": ">=3.6.1",
+    "ngx-bootstrap": "^9.0.0",
+    "popper.js": "1.16.1",
+    "rxjs": "~7.5.7",
+    "rxjs-compat": "^6.6.7",
+    "tslib": "^2.4.0",
+    "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.0",
-    "@angular/cli": "~9.1.0",
-    "@angular/compiler-cli": "~9.1.0",
-    "@angular/language-service": "~9.1.0",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.4.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.8.3"
+    "@angular-devkit/build-angular": "~14.2.6",
+    "@angular/cli": "~14.2.6",
+    "@angular/compiler-cli": "~14.2.7",
+    "@angular/language-service": "~14.2.7",
+    "@types/jasmine": "~4.3.0",
+    "@types/jasminewd2": "~2.0.10",
+    "@types/node": "^18.11.4",
+    "codelyzer": "6.0.2",
+    "jasmine-core": "~4.4.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "^6.4.1",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~10.9.1",
+    "tslint": "~6.1.3",
+    "typescript": "~4.8.4"
   }
 }
