@@ -31,7 +31,7 @@ export class LoginComponent extends BaseComponent implements OnInit {
     ngOnInit() {
 
         this.loginmodel = new LoginModel();
-        this.loginmodel.Userid = "naren.7090@gmail.com";
+        this.loginmodel.Userid = "naren.7090@testmail.com";
         this.loginmodel.Password = "1234";
         this.loginForm = this.formBuilder.group({
             'Userid': new FormControl(this.loginmodel.Userid, [Validators.required, Validators.email]),
@@ -64,10 +64,11 @@ export class LoginComponent extends BaseComponent implements OnInit {
         const result: LoginModel = Object.assign({}, this.loginForm.value);
         this.userService.userAuthentication(result.Userid, result.Password).subscribe(
             (data: any) => {
-                localStorage.setItem('userToken', data.access_token);
+                console.log(data);
+                localStorage.setItem('userToken', data.Token);
                 this.loginForm.reset();
                 this.loadMenus();
-                this.router.navigate(['/Home']);
+                this.router.navigate(['/home']);
             },
             err => {
                 this.msg = err.error.error_description;
