@@ -6,13 +6,15 @@ import { Observable } from 'rxjs/';
 import { Subscription } from 'rxjs';
 import { MessageService } from '../Service/message.service';
 import { Router } from "@angular/router";
+import {ViewEncapsulation} from '@angular/core';
+
 @Component({
     selector: 'menu-items',
     template: `<div *ngIf='userModel'>
-    <nav class='navbar navbar-inverse' >
+    <nav class='navbar navbar-expand-lg navbar-light bg-light' >
         
-            <ul class='nav navbar-nav' >
-               <li  *ngFor="let cols of userModel.routeCollection"> <a [routerLink]="[cols.key]" routerLinkActive="active" > {{ cols.keyValue }} </a></li >
+            <ul class='navbar-nav mr-auto' >
+               <li  *ngFor="let cols of userModel.routeCollection" class="nav-item active"> <a [routerLink]="[cols.key]" routerLinkActive="active" class="nav-link"  > {{ cols.keyValue }} </a></li >
             </ul>
         <ul class="nav navbar-nav navbar-right">
         <li><a href="#"><span class="glyphicon glyphicon-user"></span> {{userModel.UserName}}</a></li>
@@ -20,6 +22,7 @@ import { Router } from "@angular/router";
         </ul>
       </nav>
         </div>` ,
+         encapsulation: ViewEncapsulation.None
     // styles: ['.navbar-nav {background-color:#337AB7 !important;}']
 })
 
