@@ -3,8 +3,8 @@
 
     //public static BASE_URL="http://vmtest.australiaeast.cloudapp.azure.com/Dataapi/"
     //public static BASE_URL="http://localhost/dataapi/"
-    //public static BASE_URL = "http://192.168.2.2/dataapi/"
-    public static BASE_URL = "http://localhost:52864/"
+    public static BASE_URL = "http://192.168.1.113:5200/"
+    //public static BASE_URL = "http://localhost:62878/"
     
     //ENDPOINTS
     public static BASE_TODOLIST_ENDPOINT = "api/Todoapi/todolist";
