@@ -10,17 +10,17 @@ import { TicketComponent } from './Components/ticket.component';
 
 const routes: Routes = [
   { path: '', redirectTo: 'login', pathMatch: 'full' },
-  { path: 'Home', component: ToDoComponent },
+  { path: 'home', component: ToDoComponent },
   { path: 'login', component: LoginComponent },
   { path: 'NotAuthorized', component: NotAuthorizedComponent },
   { path: 'adduser', component: AddUserComponent },
-  { path: 'Ticket', component: TicketComponent, canActivate: [AuthGuard] },
+  { path: 'ticket', component: TicketComponent, canActivate: [AuthGuard] },
   { path: '**', component: PageNotFoundComponent, canActivate: [AuthGuard] }
 
 ];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(routes, { relativeLinkResolution: 'legacy' })],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
